mux on Signal, not Command, a new abstraction
When creating artificial events, one cannot always manipulate Command.
For example in the case of ctcp:* events, Command should remain either
PRIVMSG or NOTICE. Therefore we add a new field, Signal, which will
default to the Command but can be freely adjusted. Muxing has to be done
on this field.