permit multiple messages to terminate a coalescion group
Certain commands, like WHOIS, can also be ended by things like "no such
nick", which means we need to allow multiple messages to end a single
group.

As a side-effect, this change removes the possibility of related
messages to have different ending conditions. Where before it was
possible to have two groups "A, B, C" and "A, B" that ended on C and B
respectively, both groups now have to end on the same commands. This
should not cause any problems, because for every request, such as WHOIS,
there's always the same set of ending conditions.